+++ TestProject/core/src/Objects/BulletFormation/SlowTargetDownwardLinearBulletFormation.java
@@ -16,7 +16,7 @@ public class SlowTargetDownwardLinearBulletFormation extends BulletFormation {
     }
 
     @Override
-    public List<Ammo> shoot(Entity entity, int xShootPosition, int yShootPosition) {
+    public List<Ammo> shoot(Entity entity, float xShootPosition, float yShootPosition) {
         // Set up multiply
         float playerXPosition = player.getXPosition();
         float playerYPosition = player.getYPosition();
